rename classification flow to rater flow
